fix(checker): narrow import alias heritage fix
