+++ crates/tsz-checker/src/state/type_resolution/mod.rs
@@ -5,7 +5,6 @@ pub(crate) mod constructors;
 mod constructors_tests;
 pub mod core;
 pub(crate) mod cross_file_constructors;
-pub(crate) mod import_alias_reference;
 pub(crate) mod import_type;
 pub(crate) mod judge;
 pub(crate) mod mixin_constraints;
